# Speed up `build_mask` (~100 times)